Functionality Most of the methods to update the local database are working correctly also all the information have been stored in the datastore of google

+++ app/src/main/java/com/unal/tuapp/recapp/dialogs/FilterCategoriesDialogFragment.java
@@ -8,6 +8,7 @@ import android.support.v4.content.CursorLoader;
 import android.support.v4.content.Loader;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -47,8 +48,9 @@ public class FilterCategoriesDialogFragment extends DialogFragment implements Lo
         adapter.setOnItemClickListener(new RecycleCategoriesAdapter.OnItemClickListener(){
             @Override
             public void onItemClick(View view, long position) {
+
                 NavigationDrawer callingActivity = (NavigationDrawer) getActivity();
-                callingActivity.showSubCategoriesDialog((int)position);
+                callingActivity.showSubCategoriesDialog(position);
                 dismiss();
             }
         });
